requirements.txt
asgiref==3.8.1
astroid==3.3.8
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.0.4
black==24.10.0
cffi==1.17.1
channels==4.0.0
click==8.1.8
constantly==23.10.4
crontab==1.0.1
cryptography==44.0.0
daphne==4.0.0
dill==0.3.9
Django==5.1.4
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring-to-markdown==0.15
flake8==7.1.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==5.13.2
jedi==0.19.2
mccabe==0.7.0
msgpack==1.1.0
mypy-extensions==1.0.0
neovim==0.3.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.10.1
pylint==3.3.3
pynvim==0.5.2
pyOpenSSL==25.0.0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
pytoolconfig==1.3.1
rope==1.13.0
ruff==0.8.6
service-identity==24.2.0
setuptools==75.8.0
snowballstemmer==2.2.0
sqlparse==0.5.3
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
ujson==5.10.0
whatthepatch==1.0.7
whitenoise==6.8.2
yapf==0.43.0
zope.interface==7.2
